Remove the redirect to the /success endpoint
Currently this interrupts the login flow, and
takes the user to a `wc.org/account/success`, where
they aren't logged in and don't have a valid session.

It's not obvious how we can redirect to that page
and give them a valid session, so for now we're
going to remove the `/success` page from the flow
and send somebody straight to their account page.
This is simpler and means we're pretty close to
having a working end-to-end flow.